==> notebooks/5-assign_label_BIO.ipynb <==
import pandas as pd

chunk_size = 10000  # Adjust based on your RAM
chunks = []

for chunk in pd.read_csv('../data/processed/BioPDB_features_with_labels.csv', chunksize=chunk_size):
    processed_chunk = chunk
    chunks.append(processed_chunk)

# Combine all chunks if needed
bioPDF_df = pd.concat(chunks, ignore_index=True)

bioPDF_df['label'].value_counts()
